Upmergw stage 1

+++ core/reflection/commandLineSetter.cpp
@@ -1,4 +1,4 @@
-#include "commandLineSetter.h"
+#include "core/reflection/commandLineSetter.h"
 
 namespace corecvs {
 
@@ -17,7 +17,7 @@ void CommandLineSetter::visit<double, DoubleField>(double &field, const DoubleFi
 template <>
 void CommandLineSetter::visit<float, FloatField>(float &field, const FloatField *fieldDescriptor)
 {
-    field = getDouble(fieldDescriptor->name.name, fieldDescriptor->defaultValue);
+    field = (float)getDouble(fieldDescriptor->name.name, fieldDescriptor->defaultValue);
 }
 
 template <>
